moved exceedence utilities to mathUtils
